backend/officelink/requirements.txt
django==1.11.7
djangorestframework==3.7.3
gunicorn
dj_database_url
psycopg2
whitenoise
django-cors-headers
